Regex was capturing the request for fields beforehand.
This is super dumb but the way that we capture "custom fields" from Netsuite is by selecting 5 employees from NetSuite and then looking at those to see if theres any fields we don't have. Unfortunately, since we now expect 5000 employees in the request URL the Regex was capturing that first and not returning the correct response body in our specs. This fixes it.